fix: #43; move components into function based subfolders; refactor file org
