+++ src/Nav.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom'
-import '../App.css';
+import './App.css';
 
 class Nav extends Component {
