+++ src/posts/PostDetail.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux'
 import ArrowDownIcon from 'react-icons/lib/fa/arrow-circle-down'
 import ArrowUpIcon from 'react-icons/lib/fa/arrow-circle-up'
 import { fetchPostDetail } from '../actions'
-import Comments from './Comments'
+import Comments from '../comments/Comments'
 import { prettyTime, sortList } from '../utils'
 import '../App.css';
